+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * `CriticalInputError` is raised when the input preprocessor values can't be set, this was so far only true
   for the hmac preprocessor, but is now also applied for all other preprocessors.
+* fix `delimiter` typo in `StringSplitterRule` configuration
 
 ### Features
 ### Improvements